requirements.txt
django==3.2
python-decouple==3.4
django-heroku==0.3.1
gunicorn==20.1.0